Added filters

+++ Losliste.py
@@ -7,57 +7,60 @@ from bs4 import BeautifulSoup
 import schedule
 import threading
 
-
+errormsg = ""
 last_update = None
-fedje_id = 2353170
-#Hente data inn til database
+lock = threading.Lock()
+
 def current_pilotages():
+   payload = "__VIEWSTATE=%2FwEPDwUKLTYwMjM1ODI1OWRkihciWXOT7la0fVxgo82dzspjLffjhsDX7ILdV%2Fpg4yc%3D&ctl00%24MainContent%24PilotageDispatchDepartmentDropDown=2353122&ctl00%24MainContent%24PilotageDipatchLocationDropDown=None&ctl00%24MainContent%24ShowPilotages=Show%2BPilotages"
+   headers = {'Content-Type': "application/x-www-form-urlencoded" }
+   cest = datetime.now(tz=pytz.timezone('Europe/Oslo'))
+   head = []
+   list_data = []
+   i = 0
+   global last_update
+   global errormsg
    try:
+      #Hente data inn til database
       conn = http.client.HTTPSConnection("www.shiprep.no")
-      payload = "__VIEWSTATE=%2FwEPDwUKLTYwMjM1ODI1OWRkihciWXOT7la0fVxgo82dzspjLffjhsDX7ILdV%2Fpg4yc%3D&ctl00%24MainContent%24PilotageDispatchDepartmentDropDown=2353122&ctl00%24MainContent%24PilotageDipatchLocationDropDown=None&ctl00%24MainContent%24ShowPilotages=Show%2BPilotages"
-      headers = {'Content-Type': "application/x-www-form-urlencoded" }
       conn.request("POST", "/shiprepwebui/currentpilotages.aspx", payload, headers)
       res = conn.getresponse()
       data = res.read()
       soup = BeautifulSoup(data, 'html.parser')
+      
+      with lock:
+         for th in soup.find_all('th'):
+            head.append(th.get_text())
+            if th.get_text() == '':
+               head[i] = str(i+1)
+            i += 1
+         if len(head) < 1:
+            raise Exception("Empty response")
+         head_str = '"' + '", "'.join(head) + '"'
+         con = sqlite3.connect('motulo.db')
+         cur = con.cursor()
+         cur.execute('DROP TABLE IF EXISTS los;')
+         cur.execute(f'CREATE TABLE IF NOT EXISTS los ({head_str});')
+         
+         for tr in soup.find_all('tr'):
+            list_data.clear()
+            for td in tr.find_all('td'):
+               list_data.append(td.get_text(strip = True))
+               if len(list_data) < 10:
+                  continue
+               cur.execute('INSERT INTO los VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)', list_data)
+         con.commit()
+         cur.close()
+         con.close()
+         errormsg = ""
+         last_update = cest.strftime("%H:%M:%S")
    except Exception as exception:
       error = type(exception)
-      return("Shiprep req failed: " + str(error) + " " + str(exception))
-   head = []
-   data = []
-   i = 0
-   #Slette gammel og opprette ny table
-   try:
-      for th in soup.find_all('th'):
-         head.append(th.get_text())
-         if th.get_text() == '':
-            head[i] = str(i+1)
-         i += 1
-      head_str = '"' + '", "'.join(head) + '"'
-      con = sqlite3.connect('motulo.db')
-      cur = con.cursor()
-      cur.execute('DROP TABLE IF EXISTS los;')
-      cur.execute(f'CREATE TABLE IF NOT EXISTS los ({head_str});')
-      for tr in soup.find_all('tr'):
-         data.clear()
-         for td in tr.find_all('td'):
-            data.append(td.get_text(strip = True))
-            if len(data) < 10:
-               continue
-            cur.execute('INSERT INTO los VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)', data)
-      con.commit()
-      cur.close()
-      con.close()
-      global last_update
-      cest = datetime.now(tz=pytz.timezone('Europe/Oslo'))
-      last_update = cest.strftime("%H:%M:%S")
-   except Exception as exception:
-      error = type(exception)
-      last_update = ("Database write failure: " + str(error) + " " + str(exception))
-   
+      errormsg = "Feilet kl: " + cest.strftime("%H:%M:%S, ") + str(error) + str(exception)
+
 
 def worker():
-   schedule.every(15).seconds.do(current_pilotages)
+   schedule.every(30).seconds.do(current_pilotages)
    while True:
       schedule.run_pending()
       time.sleep(1)
@@ -68,17 +71,29 @@ def start():
    loop_thread.start()
    
 
-def des_query(des):
-   try:
-      con = sqlite3.connect('motulo.db')
-      cur = con.cursor()
-      col = '"ETA/ETD", "Ship Name", "GT", "Type", "From", "To", "Locked"'
-      query = f'SELECT {col} FROM los WHERE (LOWER("From") LIKE "{des}" OR LOWER("To") LIKE "{des}" OR LOWER("To") LIKE "sture" OR LOWER("From") LIKE "sture") AND CAST("GT" AS INTEGER) >= 20000'
-      cur.execute(query)
-      res = cur.fetchall()
-      cur.close()
-      con.close()
-      return res, last_update
-   except Exception as exception:
-      error = type(exception)
-      return("Database query failure: " + str(error) + " " + str(exception))
+def des_query(ton, des):
+    global errormsg
+    if len(des) == 0:
+       des_q = ''
+    elif type(des) == tuple and len(des) > 1:
+       des_q = f'(LOWER("From") IN {des} OR LOWER("To") IN {des}) AND'
+    elif type(des) == tuple and len(des) == 1:
+       des_q = f'(LOWER("From") LIKE "{des[0]}" OR LOWER("To") LIKE "{des[0]}") AND'
+    elif type(des) == str:
+        des_q = f'(LOWER("From") LIKE "{des}" OR LOWER("To") LIKE "{des}") AND'
+    try:
+       con = sqlite3.connect('motulo.db')
+       cur = con.cursor()
+       col = '"ETA/ETD", "Ship Name", "GT", "Type", "From", "To", "Locked"'
+       query = f'''SELECT {col} FROM los WHERE {des_q} CAST("GT" AS INTEGER) >= {ton}'''
+       cur.execute(query)
+       res = cur.fetchall()
+       cur.close()
+       con.close()
+       return res, last_update, errormsg
+    except Exception as exception:
+       error = type(exception)
+       errormsg = "Database query failure: " + str(error) + " " + str(exception)
+       res = ["feil"]
+       return res, last_update, errormsg
+
